requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pillow>=8.0.0
pywin32>=300; sys_platform == "win32"
